\name{genomicViz-package}
\alias{genomicViz-package}
\alias{genomicViz}
\docType{package}
\title{
\packageTitle{genomicViz}
}
\description{
\packageDescription{genomicViz}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{genomicViz}
\packageIndices{genomicViz}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{genomicViz}

Maintainer: \packageMaintainer{genomicViz}
}
\references{
~~ Literature or other references for background information ~~
}
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
